Clean for PHP 7 & update CI
