+++ .github/workflows/ci.yml
@@ -1,0 +1,65 @@
+name: CI
+
+on:
+  schedule:
+    - cron: 0 2 * * 1-5
+  pull_request:
+
+jobs:
+  tests:
+    name: Tests PHP ${{ matrix.php }} - Symfony ${{ matrix.symfony }} (${{ matrix.dependency }})
+    runs-on: ubuntu-latest
+    container:
+      image: php:${{ matrix.php }}
+    strategy:
+      # https://github.community/t5/GitHub-Actions/GitHub-Actions-Matrix-options-dont-work-as-documented/m-p/29558
+      matrix:
+        php:
+          - '7.2'
+          - '7.3'
+          - '7.4'
+        symfony:
+          - '3.4.*'
+          - '4.4.*'
+          - '5.0.*'
+        dependency:
+          - '--prefer-stable'
+          - '--prefer-stable --prefer-lowest'
+      fail-fast: false
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v1
+
+      - name: Install system packages
+        run: |
+          apt-get update
+          apt-get install -y --no-install-recommends \
+            unzip
+
+      - name: Disable PHP memory limit
+        run: echo 'memory_limit=-1' >> /usr/local/etc/php/php.ini
+
+      - name: Install Composer
+        run: curl -sS -o - https://raw.githubusercontent.com/composer/getcomposer.org/ba13e3fc70f1c66250d1ea7ea4911d593aa1dba5/web/installer | php -- --install-dir=/usr/local/bin --filename=composer --quiet
+
+      - name: Configure Symfony
+        run: |
+          composer global require symfony/flex
+          composer config extra.symfony.require "${{ matrix.symfony }}"
+
+      - name: Update project dependencies
+        run: composer update --no-progress --no-suggest --ansi ${{ matrix.dependency }}
+
+      - name: Run Behat tests
+        run: |
+          mkdir -p features/app/cache/test/doctrine/orm/Proxies features/app/logs
+          vendor/bin/behat
+
+      - name: Run PHPUnit tests
+        run: vendor/bin/phpunit
+
+      - name: Run PHPStan tests
+        run: vendor/bin/phpstan
+
+      - name: Run php-cs-fixer tests
+        run: vendor/bin/php-cs-fixer fix --dry-run --diff
