--- .scrutinizer.yml
@@ -1,36 +1,0 @@
-filter:
-    excluded_paths:
-        - 'features/*'
-        - 'tests/*'
-        - 'vendor/*'
-
-checks:
-    php: true
-
-coding_style:
-    php:
-        spaces:
-            before_parentheses:
-                closure_definition: true
-            around_operators:
-                concatenation: false
-
-build:
-    environment:
-        php:
-            version: 5.5.12
-    tests:
-        override:
-            -
-                command: 'vendor/bin/phpunit --coverage-clover=coverage_file'
-                coverage:
-                    file: 'coverage_file'
-                    format: 'php-clover'
-
-tools:
-    php_code_sniffer:
-        config:
-            standard: 'symfony'
-    php_cs_fixer:
-        config:
-            level: 'all'
