--- .travis.yml
@@ -1,31 +1,0 @@
-language: php
-
-sudo: false
-
-checks:
-    php:
-        code_rating: true
-        duplication: true
-
-cache:
-    directories:
-        - $HOME/.composer/cache
-
-php:
-    - 5.5
-    - 5.6
-    - 7.0
-
-matrix:
-    fast_finish: true
-    include:
-        - php: 5.5
-        - php: 5.6
-        - php: 7.0
-
-install:
-    - composer install --no-suggest --no-interaction
-
-script:
-    - vendor/bin/phpunit
-    - vendor/bin/behat
