+++ README.md
@@ -1,17 +1,17 @@
 # CoopTilleulsMigrationBundle
 
+[![Actions Status](https://github.com/coopTilleuls/CoopTilleulsMigrationBundle/workflows/CI/badge.svg)](https://github.com/coopTilleuls/CoopTilleulsMigrationBundle/actions)
+[![Packagist Version](https://img.shields.io/packagist/v/tilleuls/migration-bundle.svg?style=flat-square)](https://packagist.org/packages/tilleuls/migration-bundle)
+[![Software license](https://img.shields.io/github/license/coopTilleuls/CoopTilleulsMigrationBundle.svg?style=flat-square)](https://github.com/coopTilleuls/CoopTilleulsMigrationBundle/blob/master/LICENSE)
+
 Provides basic features for continuous migration: it implements _loaders_ & _transformers_.
 
 Feel free to contribute on it!
 
-[![Build Status](https://travis-ci.org/coopTilleuls/CoopTilleulsMigrationBundle.svg?branch=master)](https://travis-ci.org/coopTilleuls/CoopTilleulsMigrationBundle)
-[![SensioLabsInsight](https://insight.sensiolabs.com/projects/ca5b7027-d545-43dd-80fe-6bc7cd1d697b/mini.png)](https://insight.sensiolabs.com/projects/ca5b7027-d545-43dd-80fe-6bc7cd1d697b)
-[![Scrutinizer Code Quality](https://scrutinizer-ci.com/g/coopTilleuls/CoopTilleulsMigrationBundle/badges/quality-score.png?b=master)](https://scrutinizer-ci.com/g/coopTilleuls/CoopTilleulsMigrationBundle/?branch=master)
-
 ## Documentation
 
-Full documentation is available at [Resources/doc](Resources/doc).
+Full documentation is available at [Resources/doc](src/Resources/doc/getting_started.md).
 
 ## Credits
 
-Created by [Vincent Chalamon](http://vincent-chalamon.fr/) for [Les-Tilleuls.coop](https://les-tilleuls.coop/).
+Created by [Vincent Chalamon](https://github.com/vincentchalamon/) for [Les-Tilleuls.coop](https://les-tilleuls.coop/).
