+++ composer.json
@@ -16,32 +16,35 @@
         }
     ],
     "require": {
-        "php": ">=5.5",
-        "monolog/monolog": "~1.20",
-        "symfony/config": "~2.3|~3.0|~4.0",
-        "symfony/console": "~2.7|~3.0|~4.0",
-        "symfony/dependency-injection": "~3.3|~4.0",
-        "symfony/doctrine-bridge": "~2.2|~3.0|~4.0",
-        "symfony/stopwatch": "~2.3|~3.0|~4.0",
-        "symfony/http-kernel": "~2.3|~3.0|~4.0",
-        "doctrine/doctrine-bundle": "~1.6",
-        "doctrine/orm": "~2.3",
-        "doctrine/dbal": "~2.5",
-        "doctrine/annotations": "~1.2"
+        "php": ">=7.2",
+        "doctrine/annotations": "^1.2",
+        "doctrine/dbal": "^2.5",
+        "doctrine/doctrine-bundle": "^1.7.2 || ^2.0",
+        "doctrine/orm": "^2.6.3",
+        "psr/log": "^1.1",
+        "symfony/config": "^3.4 || ^4.4 || ^5.0",
+        "symfony/console": "^3.4 || ^4.4 || ^5.0",
+        "symfony/dependency-injection": "^3.4 || ^4.4 || ^5.0",
+        "symfony/doctrine-bridge": "^3.4 || ^4.4 || ^5.0",
+        "symfony/http-kernel": "^3.4 || ^4.4 || ^5.0",
+        "symfony/stopwatch": "^3.4 || ^4.4 || ^5.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.5,<5.0",
-        "behat/behat": "~3.0",
-        "behat/symfony2-extension": "~2.0",
-        "doctrine/data-fixtures": "^1.1",
-        "symfony/browser-kit": "~2.3|~3.0|~4.0",
-        "symfony/framework-bundle": "~2.8|~3.0|~4.0"
+        "behat/behat": "^3.1",
+        "doctrine/data-fixtures": "^1.2",
+        "friends-of-behat/symfony-extension": "^2.0.10@beta",
+        "friendsofphp/php-cs-fixer": "^2.15.5",
+        "phpspec/prophecy": "^1.10",
+        "phpstan/phpstan": "^0.12.17",
+        "phpunit/phpunit": "^8.0",
+        "symfony/browser-kit": "^3.4 || ^4.4 || ^5.0",
+        "symfony/framework-bundle": "^3.4 || ^4.4 || ^5.0",
+        "zendframework/zend-code": "^3.4"
     },
     "autoload": {
         "psr-4": {
-            "CoopTilleuls\\MigrationBundle\\": ""
-        },
-        "exclude-from-classmap": ["features/", "tests/"]
+            "CoopTilleuls\\MigrationBundle\\": "src/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
@@ -50,7 +53,10 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.0.x-dev"
+            "dev-master": "1.2.x-dev"
         }
+    },
+    "config": {
+        "sort-packages": true
     }
 }
