+++ features/app/LegacyBundle/Loader/UserLoader.php
@@ -1,6 +1,17 @@
 <?php
 
 /*
+ * This file is part of the MigrationBundle.
+ *
+ * (c) Vincent Chalamon <vincent@les-tilleuls.coop>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+/*
  * This file is part of the MigrationBundle package.
  *
  * (c) Vincent Chalamon <vincent@les-tilleuls.coop>
@@ -12,8 +23,8 @@
 namespace CoopTilleuls\MigrationBundle\Tests\LegacyBundle\Loader;
 
 use CoopTilleuls\MigrationBundle\Loader\AbstractLoader;
+use Doctrine\Bundle\DoctrineBundle\Registry;
 use Doctrine\DBAL\Connection;
-use Symfony\Bridge\Doctrine\RegistryInterface;
 
 /**
  * @author Vincent Chalamon <vincent@les-tilleuls.coop>
@@ -25,7 +36,7 @@ final class UserLoader extends AbstractLoader
      */
     private $connection;
 
-    public function __construct(RegistryInterface $registry, $connectionName)
+    public function __construct(Registry $registry, string $connectionName)
     {
         parent::__construct($registry, $connectionName);
         $this->connection = $registry->getConnection();
@@ -34,7 +45,7 @@ final class UserLoader extends AbstractLoader
     /**
      * {@inheritdoc}
      */
-    protected function getQuery()
+    protected function getQuery(): string
     {
         return 'SELECT * FROM user WHERE is_deleted = 0';
     }
@@ -42,7 +53,7 @@ final class UserLoader extends AbstractLoader
     /**
      * {@inheritdoc}
      */
-    protected function load(\stdClass $legacyRow)
+    protected function load(\stdClass $legacyRow): void
     {
         $legacyId = $legacyRow->id;
